% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_manage_utils.R
\name{\%&\%}
\alias{\%&\%}
\title{Useful function to easily concateate strings
Its equivalent to 'paste0' function bu nicer}
\usage{
x \%&\% y
}
\description{
Useful function to easily concateate strings
Its equivalent to 'paste0' function bu nicer
}
