% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/41_import_experiment_table.R
\name{r50x.import.experiment_table}
\alias{r50x.import.experiment_table}
\title{Parses gdx name and path and extracts experiment informations}
\usage{
r50x.import.experiment_table(
  experiment_id,
  RICE50xS3Class_selection_function,
  data_exp_table = NULL,
  last_import_time = 0,
  year_start = 1,
  year_end = r50x.SettingsS3$get_default_year_horizon()
)
}
\arguments{
\item{experiment_id}{current experiment codename}

\item{last_import_time}{sys_time of last import (to detect new files)}

\item{my_experiments}{list-structure to store experiments}
}
\description{
Parses gdx name and path and extracts experiment informations
}
