% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40_import_new_experiments_list.R
\name{r50x.import.results_as_list}
\alias{r50x.import.results_as_list}
\title{Parses gdx name and path and extracts experiment informations}
\usage{
r50x.import.results_as_list(experiment_id, my_experiments, last_import_time)
}
\arguments{
\item{experiment_id}{current experiment codename}

\item{my_experiments}{list-structure to store experiments}

\item{last_import_time}{sys_time of last import (to detect new files)}
}
\description{
Parses gdx name and path and extracts experiment informations
}
