% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20_parse_gdx_experiment.R
\name{r50x.parse.gdx_experiment}
\alias{r50x.parse.gdx_experiment}
\title{Parses gdx name and path and extracts experiment informations}
\usage{
r50x.parse.gdx_experiment(gdx_name_with_full_path)
}
\description{
Parses gdx name and path and extracts experiment informations
}
