% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/82_LEGACY.R
\name{r50x.plots.boxplot}
\alias{r50x.plots.boxplot}
\title{TODO}
\usage{
r50x.plots.boxplot(
  EXPdata,
  title,
  legend,
  yLabel,
  xLabel = "",
  LaTeX_text = FALSE,
  categories = NULL,
  colors_per_category = NULL
)
}
\description{
TODO
}
