% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/82_LEGACY.R
\name{r50x.plots.combo.map_n}
\alias{r50x.plots.combo.map_n}
\title{TODO}
\usage{
r50x.plots.combo.map_n(
  EXPdata,
  legend,
  title,
  palette = NULL,
  shape = NULL,
  min_data = NULL,
  max_data = NULL,
  columns = NULL,
  legend_symmetric = TRUE,
  legend_centre = 0,
  LaTeX_text = FALSE
)
}
\description{
TODO
}
