% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/82_LEGACY.R
\name{r50x.plots.lineplot_nty}
\alias{r50x.plots.lineplot_nty}
\title{TODO}
\usage{
r50x.plots.lineplot_nty(
  EXPdata,
  title = "RICEx nty lineplot",
  y_label = "Value",
  x_label = "Year",
  legend_columns = NULL,
  region_code = "ed57",
  LaTeX_text = FALSE,
  no_legend = FALSE
)
}
\description{
TODO
}
