% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/82_LEGACY.R
\name{r50x.plots.lineplot_ty}
\alias{r50x.plots.lineplot_ty}
\title{TODO}
\usage{
r50x.plots.lineplot_ty(
  EXPdata,
  title,
  legend,
  y_label,
  x_label = "Year",
  LaTeX_text = FALSE,
  categories = NULL,
  colors_per_category = NULL,
  legend_columns = NULL,
  yLabel = NULL,
  xLabel = NULL,
  show_confidence = FALSE
)
}
\description{
TODO
}
