% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/82_LEGACY.R
\name{r50x.plots.map_n}
\alias{r50x.plots.map_n}
\title{TODO}
\usage{
r50x.plots.map_n(
  data,
  legend,
  title,
  palette,
  shape,
  min_data,
  max_data,
  centre_data,
  LaTeX_text = FALSE
)
}
\description{
TODO
}
