% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/64_plots_utils_map_shpdata_selector.R
\docType{data}
\name{r50x.plots.utils.map_shapedata_selector}
\alias{r50x.plots.utils.map_shapedata_selector}
\title{Returns .shp datafile for ggplot maps for specific regional disaggregations}
\format{An object of class \code{list} of length 4.}
\usage{
r50x.plots.utils.map_shapedata_selector
}
\description{
Returns .shp datafile for ggplot maps for specific regional disaggregations
}
\keyword{datasets}
