% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_model_settings.R
\name{r50x.Settings.create_settings_environment}
\alias{r50x.Settings.create_settings_environment}
\title{Function to update model results folder}
\usage{
r50x.Settings.create_settings_environment()
}
\description{
Function to update model results folder
}
