% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_model_settings.R
\name{r50x.Settings.set_results_folder}
\alias{r50x.Settings.set_results_folder}
\title{Function to update model results folder}
\usage{
r50x.Settings.set_results_folder(path)
}
\arguments{
\item{path}{Full path to your RICE50x results folder}
}
\description{
Function to update model results folder
}
