% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_model_settings.R
\docType{data}
\name{r50x.SettingsS3}
\alias{r50x.SettingsS3}
\title{Function to update model results folder}
\format{An object of class \code{list} (inherits from \code{ModelEnvClass}) of length 6.}
\usage{
r50x.SettingsS3
}
\description{
Function to update model results folder
}
\keyword{datasets}
