% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/63_plots_utils.R
\name{r50x.utils.listify_EXPdataframe}
\alias{r50x.utils.listify_EXPdataframe}
\title{function to generate list from dataframe, based on specific value}
\usage{
r50x.utils.listify_EXPdataframe(EXPdataframe, column_to_list)
}
\description{
function to generate list from dataframe, based on specific value
}
