% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_RICE50x_S3class.R
\name{RICE50xS3Class}
\alias{RICE50xS3Class}
\title{S3 CLASS containing useful getter for all needed parameters}
\usage{
RICE50xS3Class(gdx_file_and_path)
}
\arguments{
\item{gdx_file_and_path}{string path to gdx file associated}
}
\description{
S3 CLASS containing useful getter for all needed parameters
}
