% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/41_import_experiment_table.R
\name{RICEx.import.experiment_table}
\alias{RICEx.import.experiment_table}
\title{Parses gdx name and path and extracts experiment informations}
\usage{
RICEx.import.experiment_table(
  experiment_id,
  RICExclass_selection_function,
  data_exp_table = NULL,
  last_import_time = 0
)
}
\arguments{
\item{experiment_id}{current experiment codename}

\item{last_import_time}{sys_time of last import (to detect new files)}

\item{my_experiments}{list-structure to store experiments}
}
\description{
Parses gdx name and path and extracts experiment informations
}
