% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_manage_utils.R
\name{WORLDaggr_ntySUMty}
\alias{WORLDaggr_ntySUMty}
\title{Function to aggregate nty dataframes into world ty ones}
\usage{
WORLDaggr_ntySUMty(DF_nty)
}
\description{
Function to aggregate nty dataframes into world ty ones
}
